// src/a2aclient/types.ts
export type Part = TextPart | FilePart | DataPart;
/**
 * Mirrors the OpenAPI Security Scheme Object
 * (https://swagger.io/specification/#security-scheme-object)
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "SecurityScheme".
 */
export type SecurityScheme =
  | APIKeySecurityScheme
  | HTTPAuthSecurityScheme
  | OAuth2SecurityScheme
  | OpenIdConnectSecurityScheme;
// ...
export interface Message {
  /**
   * The context the message is associated with
   */
  contextId?: string;
  /**
   * The URIs of extensions that are present or contributed to this Message.
   */
  extensions?: string[];
  /**
   * Event type
   */
  kind: "message";
  /**
   * Identifier created by the message creator
   */
  messageId: string;
  /**
   * Extension metadata.
   */
  metadata?: {
    [k: string]: unknown;
  };
  /**
   * Message content
   */
  parts: Part[];
  /**
   * List of tasks referenced as context by this message.
   */
  referenceTaskIds?: string[];
  /**
   * Message sender's role
   */
  role: "agent" | "user";
  /**
   * Identifier of task the message is related to
   */
  taskId?: string;
}
/**
 * Represents a text segment within parts.
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "TextPart".
 */
export interface TextPart {
  /**
   * Part type - text for TextParts
   */
  kind: "text";
  /**
   * Optional metadata associated with the part.
   */
  metadata?: {
    [k: string]: unknown;
  };
  /**
   * Text content
   */
  text: string;
}
/**
 * Represents a File segment within parts.
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "FilePart".
 */
export interface FilePart {
  /**
   * File content either as url or bytes
   */
  file: FileWithBytes | FileWithUri;
  /**
   * Part type - file for FileParts
   */
  kind: "file";
  /**
   * Optional metadata associated with the part.
   */
  metadata?: {
    [k: string]: unknown;
  };
}
// ...
export interface DataPart {
  /**
   * Structured data content
   */
  data: {
    [k: string]: unknown;
  };
  /**
   * Part type - data for DataParts
   */
  kind: "data";
  /**
   * Optional metadata associated with the part.
   */
  metadata?: {
    [k: string]: unknown;
  };
}
// ...
export interface APIKeySecurityScheme {
  /**
   * Description of this security scheme.
   */
  description?: string;
  /**
   * The location of the API key. Valid values are "query", "header", or "cookie".
   */
  in: "cookie" | "header" | "query";
  /**
   * The name of the header, query or cookie parameter to be used.
   */
  name: string;
  type: "apiKey";
}
// ...
export interface AgentExtension {
  /**
   * A description of how this agent uses this extension.
   */
  description?: string;
  /**
   * Optional configuration for the extension.
   */
  params?: {
    [k: string]: unknown;
  };
  /**
   * Whether the client must follow specific requirements of the extension.
   */
  required?: boolean;
  /**
   * The URI of the extension.
   */
  uri: string;
}
/**
 * An AgentCard conveys key information:
 * - Overall details (version, name, description, uses)
 * - Skills: A set of capabilities the agent can perform
 * - Default modalities/content types supported by the agent.
 * - Authentication requirements
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "AgentCard".
 */
export interface AgentCard {
  capabilities: AgentCapabilities1;
  /**
   * The set of interaction modes that the agent supports across all skills. This can be overridden per-skill.
   * Supported media types for input.
   */
  defaultInputModes: string[];
  /**
   * Supported media types for output.
   */
  defaultOutputModes: string[];
  /**
   * A human-readable description of the agent. Used to assist users and
   * other agents in understanding what the agent can do.
   */
  description: string;
  /**
   * A URL to documentation for the agent.
   */
  documentationUrl?: string;
  /**
   * A URL to an icon for the agent.
   */
  iconUrl?: string;
  /**
   * Human readable name of the agent.
   */
  name: string;
  provider?: AgentProvider;
  /**
   * Security requirements for contacting the agent.
   */
  security?: {
    [k: string]: string[];
  }[];
  /**
   * Security scheme details used for authenticating with this agent.
   */
  securitySchemes?: {
    [k: string]: SecurityScheme;
  };
  /**
   * Skills are a unit of capability that an agent can perform.
   */
  skills: AgentSkill[];
  /**
   * true if the agent supports providing an extended agent card when the user is authenticated.
   * Defaults to false if not specified.
   */
  supportsAuthenticatedExtendedCard?: boolean;
  /**
   * A URL to the address the agent is hosted at.
   */
  url: string;
  /**
   * The version of the agent - format is up to the provider.
   */
  version: string;
}
/**
 * Optional capabilities supported by the agent.
 */
export interface AgentCapabilities1 {
  /**
   * extensions supported by this agent.
   */
  extensions?: AgentExtension[];
  /**
   * true if the agent can notify updates to client.
   */
  pushNotifications?: boolean;
  /**
   * true if the agent exposes status change history for tasks.
   */
  stateTransitionHistory?: boolean;
  /**
   * true if the agent supports SSE.
   */
  streaming?: boolean;
}
/**
 * The service provider of the agent
 */
export interface AgentProvider {
  /**
   * Agent provider's organization name.
   */
  organization: string;
  /**
   * Agent provider's URL.
   */
  url: string;
}
/**
 * HTTP Authentication security scheme.
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "HTTPAuthSecurityScheme".
 */
export interface HTTPAuthSecurityScheme {
  /**
   * A hint to the client to identify how the bearer token is formatted. Bearer tokens are usually
   * generated by an authorization server, so this information is primarily for documentation
   * purposes.
   */
  bearerFormat?: string;
  /**
   * Description of this security scheme.
   */
  description?: string;
  /**
   * The name of the HTTP Authentication scheme to be used in the Authorization header as defined
   * in RFC7235. The values used SHOULD be registered in the IANA Authentication Scheme registry.
   * The value is case-insensitive, as defined in RFC7235.
   */
  scheme: string;
  type: "http";
}
/**
 * OAuth2.0 security scheme configuration.
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "OAuth2SecurityScheme".
 */
export interface OAuth2SecurityScheme {
  /**
   * Description of this security scheme.
   */
  description?: string;
  flows: OAuthFlows;
  type: "oauth2";
}
// ...
export interface OpenIdConnectSecurityScheme {
  /**
   * Description of this security scheme.
   */
  description?: string;
  /**
   * Well-known URL to discover the [[OpenID-Connect-Discovery]] provider metadata.
   */
  openIdConnectUrl: string;
  type: "openIdConnect";
}
/**
 * Represents a unit of capability that an agent can perform.
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "AgentSkill".
 */
export interface AgentSkill {
  /**
   * Description of the skill - will be used by the client or a human
   * as a hint to understand what the skill does.
   */
  description: string;
  /**
   * The set of example scenarios that the skill can perform.
   * Will be used by the client as a hint to understand how the skill can be used.
   */
  examples?: string[];
  /**
   * Unique identifier for the agent's skill.
   */
  id: string;
  /**
   * The set of interaction modes that the skill supports
   * (if different than the default).
   * Supported media types for input.
   */
  inputModes?: string[];
  /**
   * Human readable name of the skill.
   */
  name: string;
  /**
   * Supported media types for output.
   */
  outputModes?: string[];
  /**
   * Set of tagwords describing classes of capabilities for this specific skill.
   */
  tags: string[];
}